tests

+++ server/src/test/java/com/cdyweb/tc/TestDecode.java
@@ -15,22 +15,22 @@ public class TestDecode {
     byte[] buf=new byte[2];
     buf[0]=25;
     buf[1]=0x38;
-    float f1 = BV4235.decode_temperature(buf);
+    double f1 = BV4235.decode_temperature(buf);
     System.out.println(f1);
     
     buf[0]=25;
     buf[1]=(byte)0xF0;
-    float f2 = BV4235.decode_temperature(buf);
+    double f2 = BV4235.decode_temperature(buf);
     System.out.println(f2);
 
     buf[0]=-25;
     buf[1]=(byte)0xF0;
-    float f3 = BV4235.decode_temperature(buf);
+    double f3 = BV4235.decode_temperature(buf);
     System.out.println(f3);
 
     buf[0]=-25;
     buf[1]=0x38;
-    float f4 = BV4235.decode_temperature(buf);
+    double f4 = BV4235.decode_temperature(buf);
     System.out.println(f4);
   }
 }
